lando: packaging (bug 1866325)
- add pyproject.toml
- move project files to main lando module
- update settings to reflect package structure
- update Dockerfile to use package
