+++ .gitignore
@@ -1,3 +1,5 @@
 env
 staticfiles
 .env
+__pycache__
+*.egg-info
