+++ src/lando/utils/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class UtilsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "utils"
+    name = "lando.utils"
